moved projector to its own file.
